requirements.txt
# Core dependencies
overrides>=7.4.0
ml-collections>=0.1.1
tensorflow==2.15.0
tensorflow-hub>=0.16.0
tensorflow-datasets>=4.9.0
tensorflow-probability>=0.23.0
tensorboard>=2.15.0
torch>=2.1.0
torchvision>=0.16.0
torchmetrics>=1.2.0
pytorch-fid>=0.3.0
absl-py>=2.0.0
ninja
scipy>=1.10.0
natsort
imageio
numpy>=1.24.0,<2.0.0
pillow>=10.0.0
matplotlib>=3.7.0
h5py>=3.8.0
wrapt>=1.14.0

# Optional: JAX support (commented out by default)
# jax[cuda12_pip]>=0.4.20  # For CUDA 12.x
# jaxlib>=0.4.20
